Add shebang to gpt4-script
